componentsディレクトリを_componentsに名称変更
