+++ src/app/[facultyCategoryPathName]/_components/FacultyList.tsx
