+++ src/app/[facultyCategoryPathName]/_components/__tests__/FacultyList.test.tsx
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom";
 import FacultySearchParams from "@/types/FacultySearchParams";
 import { getPostOffices } from "@/fixtures/post_offices";
 import * as fetchFacultiesResultFetchers from "@/lib/fetchFaculties";
-import FacultyList from "@/app/[facultyCategoryPathName]/components/FacultyList";
+import FacultyList from "@/app/[facultyCategoryPathName]/_components/FacultyList";
 
 jest.mock("src/lib/fetchFaculties");
