+++ src/app/[facultyCategoryPathName]/_components/__tests__/FacultySearchForm.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom";
-import VillageSearchForm from "@/app/components/VillageSearchForm";
 import selectEvent from "react-select-event";
+import FacultySearchForm from "@/app/[facultyCategoryPathName]/_components/FacultySearchForm";
 
 const user = userEvent.setup();
 
@@ -20,13 +20,13 @@ HTMLDialogElement.prototype.showModal = jest.fn(function mock(
   this.open = true;
 });
 
-describe("VillageSearchForm", () => {
+describe("FacultySearchForm", () => {
   beforeEach(() => {
     mockFn.mockClear();
   });
 
   it("地域を選択しオプションはデフォルト値で検索する", async () => {
-    render(<VillageSearchForm />);
+    render(<FacultySearchForm facultyCategoryPathName="post_office" />);
     const button = screen.getByRole("button", { name: "探索" });
     expect(button).toBeDisabled();
 
@@ -36,36 +36,22 @@ describe("VillageSearchForm", () => {
 
     const params = new URLSearchParams({
       area: "aomori",
-      populationLowerLimit: "1",
-      populationUpperLimit: "10000",
       islandSetting: "exclude_islands",
       keywords: "",
       page: "1",
     });
 
     await user.click(button);
-    expect(mockFn).toHaveBeenCalledWith(`/?${params.toString()}`);
+    expect(mockFn).toHaveBeenCalledWith(`/post_office/?${params.toString()}`);
   });
 
   it("地域とオプションを選択して検索する", async () => {
-    render(<VillageSearchForm />);
+    render(<FacultySearchForm facultyCategoryPathName="post_office" />);
 
     const areaSelectBox = screen.getByRole("combobox");
     await selectEvent.select(areaSelectBox, "青森県");
 
     await user.click(screen.getByRole("button", { name: "詳細条件" }));
-
-    expect(screen.getByRole("group", { name: "人口" })).toBeInTheDocument();
-    const populationLowerLimitTextbox = screen.getByRole("spinbutton", {
-      name: "最小：",
-    });
-    await user.clear(populationLowerLimitTextbox);
-    await user.type(populationLowerLimitTextbox, "200");
-    const populationUpperLimitTextbox = screen.getByRole("spinbutton", {
-      name: "最大：",
-    });
-    await user.clear(populationUpperLimitTextbox);
-    await user.type(populationUpperLimitTextbox, "500");
 
     expect(screen.getByRole("group", { name: "離島設定" })).toBeInTheDocument();
     await user.click(screen.getByLabelText("離島のみ"));
@@ -79,36 +65,22 @@ describe("VillageSearchForm", () => {
 
     const params = new URLSearchParams({
       area: "aomori",
-      populationLowerLimit: "200",
-      populationUpperLimit: "500",
       islandSetting: "only_islands",
       keywords: "佐井村",
       page: "1",
     });
 
     await user.click(screen.getByRole("button", { name: "探索" }));
-    expect(mockFn).toHaveBeenCalledWith(`/?${params.toString()}`);
+    expect(mockFn).toHaveBeenCalledWith(`/post_office/?${params.toString()}`);
   });
 
   it("地域とオプションを選択した後、デフォルト値に戻して検索する", async () => {
-    render(<VillageSearchForm />);
+    render(<FacultySearchForm facultyCategoryPathName="post_office" />);
 
     const areaSelectBox = screen.getByRole("combobox");
     await selectEvent.select(areaSelectBox, "青森県");
 
     await user.click(screen.getByRole("button", { name: "詳細条件" }));
-
-    expect(screen.getByRole("group", { name: "人口" })).toBeInTheDocument();
-    const populationLowerLimitTextbox = screen.getByRole("spinbutton", {
-      name: "最小：",
-    });
-    await user.clear(populationLowerLimitTextbox);
-    await user.type(populationLowerLimitTextbox, "200");
-    const populationUpperLimitTextbox = screen.getByRole("spinbutton", {
-      name: "最大：",
-    });
-    await user.clear(populationUpperLimitTextbox);
-    await user.type(populationUpperLimitTextbox, "500");
 
     expect(screen.getByRole("group", { name: "離島設定" })).toBeInTheDocument();
     await user.click(screen.getByLabelText("離島のみ"));
@@ -125,23 +97,20 @@ describe("VillageSearchForm", () => {
 
     const params = new URLSearchParams({
       area: "aomori",
-      populationLowerLimit: "1",
-      populationUpperLimit: "10000",
       islandSetting: "exclude_islands",
       keywords: "",
       page: "1",
     });
 
     await user.click(screen.getByRole("button", { name: "探索" }));
-    expect(mockFn).toHaveBeenCalledWith(`/?${params.toString()}`);
+    expect(mockFn).toHaveBeenCalledWith(`/post_office/?${params.toString()}`);
   });
 
   it("propsがフォームの初期値に反映されている", async () => {
     render(
-      <VillageSearchForm
+      <FacultySearchForm
+        facultyCategoryPathName="post_office"
         inputArea={{ enName: "aomori", jpName: "青森県" }}
-        inputPopulationLowerLimit="200"
-        inputPopulationUpperLimit="500"
         inputIslandSetting={{
           jpName: "離島のみ",
           enName: "only_islands",
@@ -154,16 +123,6 @@ describe("VillageSearchForm", () => {
 
     await user.click(screen.getByRole("button", { name: "詳細条件" }));
 
-    expect(screen.getByRole("group", { name: "人口" })).toBeInTheDocument();
-    const populationLowerLimitTextbox = screen.getByRole("spinbutton", {
-      name: "最小：",
-    });
-    expect(populationLowerLimitTextbox).toHaveValue(200);
-    const populationUpperLimitTextbox = screen.getByRole("spinbutton", {
-      name: "最大：",
-    });
-    expect(populationUpperLimitTextbox).toHaveValue(500);
-
     expect(screen.getByRole("group", { name: "離島設定" })).toBeInTheDocument();
     expect(screen.getByLabelText("離島のみ")).toBeChecked();
 
@@ -175,14 +134,12 @@ describe("VillageSearchForm", () => {
 
     const params = new URLSearchParams({
       area: "aomori",
-      populationLowerLimit: "200",
-      populationUpperLimit: "500",
       islandSetting: "only_islands",
       keywords: "佐井村",
       page: "1",
     });
 
     await user.click(screen.getByRole("button", { name: "探索" }));
-    expect(mockFn).toHaveBeenCalledWith(`/?${params.toString()}`);
+    expect(mockFn).toHaveBeenCalledWith(`/post_office/?${params.toString()}`);
   });
 });
