+++ src/app/_components/__tests__/VillageList.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import type VillageSearchParams from "@/types/VillageSearchParams";
 import { getVillages } from "@/fixtures/villages";
-import VillageList from "@/app/components/VillageList";
+import VillageList from "@/app/_components/VillageList";
 import * as FetchVillagesResultFetchers from "@/lib/fetchVillages";
 
 jest.mock("src/lib/fetchVillages");
