+++ src/app/fortune/[facultyCategoryPathName]/_components/FacultyFortuneModal.tsx
