+++ src/app/fortune/[facultyCategoryPathName]/page.tsx
@@ -1,4 +1,4 @@
-import FacultyFortuneModal from "@/app/fortune/[facultyCategoryPathName]/components/FacultyFortuneModal";
+import FacultyFortuneModal from "@/app/fortune/[facultyCategoryPathName]/_components/FacultyFortuneModal";
 import {
   facultyCategoryPathNames,
   getFacultyCategoryFromPathName,
