+++ src/app/fortune/_components/__tests__/VillageFortuneModal.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom";
 import { village } from "@/fixtures/villages";
-import VillageFortuneModal from "@/app/fortune/components/VillageFortuneModal";
+import VillageFortuneModal from "@/app/fortune/_components/VillageFortuneModal";
 import * as VillageFortuneResultFetchers from "@/lib/fetchVillageFortuneResult";
 
 const user = userEvent.setup();
